chore: Rename folder to expo-passkeys
