+++ packages/clerk-js/bundlewatch.config.json
@@ -5,7 +5,7 @@
     { "path": "./dist/clerk.headless.js", "maxSize": "48kB" },
     { "path": "./dist/ui-common*.js", "maxSize": "87KB" },
     { "path": "./dist/vendors*.js", "maxSize": "70KB" },
-    { "path": "./dist/coinbase*.js", "maxSize": "59KB" },
+    { "path": "./dist/coinbase*.js", "maxSize": "58KB" },
     { "path": "./dist/createorganization*.js", "maxSize": "5KB" },
     { "path": "./dist/impersonationfab*.js", "maxSize": "5KB" },
     { "path": "./dist/organizationprofile*.js", "maxSize": "12KB" },
