+++ packages/expo-passkeys/app.json
@@ -1,7 +1,7 @@
 {
   "expo": {
-    "name": "clerk-expo-passkeys",
-    "slug": "clerk-expo-passkeys",
+    "name": "expo-passkeys",
+    "slug": "expo-passkeys",
     "version": "0.1.0",
     "description": "Passkeys for expo built by clerk team",
     "sdkVersion": "51.0.0",
