+++ packages/expo-passkeys/example/assets/adaptive-icon.png
