+++ packages/expo-passkeys/example/assets/icon.png
