+++ packages/expo-passkeys/example/assets/splash.png
