+++ packages/expo-passkeys/ios/ClerkExpoPasskeys.podspec
@@ -12,7 +12,7 @@ Pod::Spec.new do |s|
   s.homepage       = package['homepage']
   s.platforms      = { :ios => '13.4', :tvos => '13.4' }
   s.swift_version  = '5.4'
-  s.source         = { git: 'https://github.com/AlexNti/clerk-expo-passkeys' }
+  s.source         = { git: 'https://github.com/clerk/javascript/tree/main/packages/expo-passkeys' }
   s.static_framework = true
 
   s.dependency 'ExpoModulesCore'
